requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
aiofiles==23.2.1

sentence-transformers==3.0.1
faiss-cpu==1.8.0

pypdf==4.2.0
python-docx==1.1.2
openpyxl==3.1.5

numpy==1.26.4
pydantic==2.7.4
python-dotenv==1.0.1

requests==2.32.3